Revert "repo: update curtime to double"
This reverts commit 52b188aad230b755883faf1a08406bb403732ef2.

+++ game/client/c_baseplayer.h
@@ -310,7 +310,7 @@ public:
 
 	virtual void			SetAnimation( PLAYER_ANIM playerAnim );
 
-	double					GetTimeBase( void ) const;
+	float					GetTimeBase( void ) const;
 	float					GetFinalPredictedTime() const;
 
 	bool					IsInVGuiInputMode() const;
